attendanceManager/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2
pydantic
anthropic
openai
google-generativeai
tenacity
google-genai